Add avsregistry utils to chain_clients file in testutils folder
