+++ testing/testing-utils/Cargo.toml
@@ -14,6 +14,7 @@ workspace = true
 alloy.workspace = true
 eigen-common.workspace = true
 eigen-utils.workspace = true
+eigen-crypto-bls.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 testcontainers.workspace = true
